Add iterative wawa method and correct misspell in test_model. (iterative wawa test are not good because of randomness of MV)
