+++ peerannot/models/__init__.py
@@ -6,7 +6,10 @@ from .aggregation.GLAD import GLAD
 from .aggregation.WDS import WDS
 from .aggregation.twothird import TwoThird
 from .aggregation.plantnet import PlantNet
+
 from .aggregation.Wawa import Wawa
+from .aggregation.IterativeWawa import IterativeWawa
+
 from .identification.WAUM_perworker import WAUM_perworker
 from .identification.WAUM import WAUM
 from .identification.AUM import AUM
@@ -26,6 +29,7 @@ agg_strategies = {
     "WDS": WDS,
     "PlantNet": PlantNet,
     "TwoThird": TwoThird,
+    "IterativeWawa": IterativeWawa,
     "Wawa": Wawa,
 }
