+++ tests/test_models.py
@@ -4,6 +4,7 @@ import json
 import shutil
 
 dir_toydata = Path(__file__).parents[1] / "datasets" / "toy-data"
+dir_music = Path(__file__).parents[1] / "datasets" / "music"
 
 
 def load_json(path):
@@ -12,6 +13,7 @@ def load_json(path):
 
 
 ANSWERS = load_json(dir_toydata / "answers.json")
+ANSWERS_MUSIC = load_json(dir_music / "answers.json")
 
 
 def test_mv():
@@ -19,7 +21,7 @@ def test_mv():
     from peerannot.models import MV
 
     for sparse_ in sparse:
-        mv = MV(ANSWERS, n_classses=2, sparse=sparse_)
+        mv = MV(ANSWERS, n_classes=2, sparse=sparse_)
         y = mv.get_answers()
         expected = np.array([1, 0, 1])
         assert all([e == y_ for e, y_ in zip(expected, y)])
@@ -55,12 +57,734 @@ def test_wawa():
 
     sparse = [True, False]
     for sparse_ in sparse:
-        wawa = Wawa(ANSWERS, n_classses=2, n_workers=4, sparse=sparse_)
+        wawa = Wawa(ANSWERS, n_classes=2, n_workers=4, sparse=sparse_)
         wawa.run()
         y = wawa.get_answers()
         expected = np.array([1, 0, 1])
         assert np.isclose((wawa.worker_score - np.array([1 / 2, 1, 1, 1 / 2])).sum(), 0)
         assert all([e == y_ for e, y_ in zip(expected, y)])
+        # assert True == False
+
+
+def test_iterative_wawa():
+    from peerannot.models import IterativeWawa
+
+    sparse = [True, False]
+    for sparse_ in sparse:
+        wawa = IterativeWawa(ANSWERS_MUSIC, n_classes=10, n_workers=44, sparse=sparse_)
+        wawa.run(1)
+        y1 = wawa.get_answers()
+        # print(y1)
+        wawa.run(20)
+        y10 = wawa.get_answers()
+        print(y10)
+        expected = np.array(
+            [
+                7,
+                7,
+                4,
+                7,
+                8,
+                7,
+                2,
+                2,
+                8,
+                3,
+                0,
+                9,
+                7,
+                2,
+                2,
+                0,
+                5,
+                2,
+                8,
+                1,
+                2,
+                1,
+                7,
+                8,
+                5,
+                5,
+                1,
+                2,
+                5,
+                8,
+                2,
+                2,
+                9,
+                4,
+                4,
+                9,
+                2,
+                3,
+                2,
+                3,
+                6,
+                1,
+                8,
+                8,
+                7,
+                7,
+                3,
+                0,
+                2,
+                2,
+                7,
+                8,
+                0,
+                7,
+                5,
+                5,
+                9,
+                9,
+                6,
+                2,
+                5,
+                2,
+                7,
+                2,
+                5,
+                8,
+                7,
+                7,
+                7,
+                1,
+                5,
+                3,
+                7,
+                9,
+                0,
+                6,
+                0,
+                7,
+                3,
+                8,
+                1,
+                5,
+                1,
+                5,
+                3,
+                5,
+                3,
+                2,
+                9,
+                7,
+                8,
+                4,
+                8,
+                3,
+                7,
+                0,
+                8,
+                0,
+                8,
+                0,
+                1,
+                9,
+                4,
+                3,
+                1,
+                9,
+                2,
+                3,
+                0,
+                6,
+                4,
+                9,
+                2,
+                7,
+                5,
+                6,
+                7,
+                5,
+                9,
+                4,
+                9,
+                7,
+                7,
+                7,
+                9,
+                7,
+                0,
+                7,
+                6,
+                7,
+                7,
+                2,
+                4,
+                1,
+                2,
+                9,
+                8,
+                6,
+                1,
+                7,
+                9,
+                2,
+                9,
+                1,
+                7,
+                5,
+                8,
+                5,
+                2,
+                2,
+                9,
+                7,
+                2,
+                4,
+                9,
+                5,
+                4,
+                4,
+                4,
+                9,
+                8,
+                6,
+                5,
+                0,
+                1,
+                8,
+                1,
+                3,
+                9,
+                8,
+                9,
+                9,
+                5,
+                1,
+                1,
+                2,
+                7,
+                0,
+                9,
+                2,
+                2,
+                9,
+                7,
+                3,
+                1,
+                2,
+                4,
+                7,
+                2,
+                1,
+                2,
+                5,
+                5,
+                0,
+                4,
+                9,
+                2,
+                3,
+                5,
+                0,
+                7,
+                1,
+                4,
+                9,
+                2,
+                6,
+                4,
+                0,
+                3,
+                6,
+                3,
+                5,
+                1,
+                8,
+                7,
+                1,
+                9,
+                6,
+                6,
+                2,
+                1,
+                7,
+                3,
+                6,
+                2,
+                5,
+                1,
+                3,
+                9,
+                2,
+                3,
+                8,
+                3,
+                5,
+                1,
+                2,
+                7,
+                0,
+                3,
+                7,
+                1,
+                1,
+                1,
+                4,
+                7,
+                9,
+                9,
+                5,
+                6,
+                8,
+                5,
+                4,
+                0,
+                5,
+                3,
+                1,
+                1,
+                7,
+                9,
+                9,
+                8,
+                9,
+                1,
+                7,
+                5,
+                8,
+                8,
+                2,
+                1,
+                4,
+                0,
+                9,
+                7,
+                7,
+                5,
+                2,
+                4,
+                1,
+                7,
+                7,
+                0,
+                3,
+                0,
+                1,
+                1,
+                0,
+                0,
+                9,
+                0,
+                3,
+                2,
+                7,
+                5,
+                7,
+                7,
+                9,
+                4,
+                2,
+                8,
+                2,
+                1,
+                1,
+                8,
+                9,
+                9,
+                9,
+                2,
+                4,
+                1,
+                5,
+                7,
+                5,
+                9,
+                8,
+                7,
+                5,
+                7,
+                8,
+                7,
+                1,
+                7,
+                1,
+                1,
+                7,
+                1,
+                9,
+                5,
+                4,
+                8,
+                4,
+                8,
+                6,
+                3,
+                1,
+                5,
+                0,
+                9,
+                8,
+                8,
+                4,
+                2,
+                0,
+                9,
+                9,
+                7,
+                1,
+                0,
+                0,
+                3,
+                4,
+                5,
+                8,
+                1,
+                7,
+                5,
+                8,
+                4,
+                2,
+                5,
+                2,
+                7,
+                9,
+                2,
+                5,
+                2,
+                8,
+                8,
+                8,
+                0,
+                1,
+                4,
+                9,
+                6,
+                5,
+                4,
+                6,
+                9,
+                8,
+                0,
+                5,
+                6,
+                7,
+                4,
+                1,
+                5,
+                4,
+                8,
+                4,
+                7,
+                8,
+                3,
+                6,
+                7,
+                4,
+                6,
+                0,
+                6,
+                8,
+                8,
+                7,
+                3,
+                7,
+                7,
+                9,
+                7,
+                8,
+                2,
+                2,
+                6,
+                2,
+                9,
+                0,
+                0,
+                4,
+                1,
+                4,
+                5,
+                2,
+                9,
+                8,
+                7,
+                8,
+                7,
+                0,
+                7,
+                6,
+                7,
+                7,
+                6,
+                7,
+                4,
+                6,
+                5,
+                8,
+                5,
+                4,
+                1,
+                4,
+                7,
+                4,
+                0,
+                3,
+                9,
+                7,
+                5,
+                2,
+                2,
+                9,
+                8,
+                2,
+                7,
+                1,
+                3,
+                4,
+                4,
+                0,
+                9,
+                4,
+                7,
+                2,
+                2,
+                9,
+                4,
+                5,
+                2,
+                8,
+                6,
+                8,
+                3,
+                3,
+                7,
+                5,
+                9,
+                5,
+                3,
+                3,
+                0,
+                8,
+                5,
+                1,
+                3,
+                9,
+                7,
+                3,
+                2,
+                6,
+                9,
+                3,
+                1,
+                7,
+                3,
+                0,
+                0,
+                5,
+                9,
+                9,
+                4,
+                7,
+                5,
+                3,
+                5,
+                0,
+                2,
+                7,
+                2,
+                4,
+                9,
+                5,
+                5,
+                2,
+                5,
+                4,
+                7,
+                9,
+                7,
+                5,
+                4,
+                1,
+                3,
+                5,
+                7,
+                2,
+                1,
+                7,
+                7,
+                7,
+                7,
+                0,
+                0,
+                2,
+                7,
+                6,
+                2,
+                4,
+                8,
+                3,
+                1,
+                9,
+                7,
+                0,
+                3,
+                3,
+                8,
+                8,
+                7,
+                9,
+                2,
+                8,
+                7,
+                1,
+                7,
+                7,
+                2,
+                7,
+                8,
+                2,
+                0,
+                4,
+                9,
+                4,
+                9,
+                1,
+                5,
+                4,
+                0,
+                4,
+                0,
+                1,
+                7,
+                2,
+                8,
+                6,
+                4,
+                1,
+                6,
+                0,
+                7,
+                1,
+                7,
+                9,
+                8,
+                7,
+                7,
+                4,
+                9,
+                4,
+                5,
+                2,
+                5,
+                9,
+                4,
+                5,
+                4,
+                5,
+                0,
+                4,
+                9,
+                6,
+                7,
+                1,
+                1,
+                7,
+                0,
+                7,
+                4,
+                8,
+                2,
+                3,
+                2,
+                3,
+                9,
+                6,
+                9,
+                1,
+                7,
+                1,
+                7,
+                8,
+                9,
+                5,
+                4,
+                3,
+                5,
+                7,
+                7,
+                9,
+                9,
+                7,
+                8,
+                4,
+                8,
+                9,
+                5,
+                3,
+                9,
+                3,
+                5,
+                7,
+                1,
+                4,
+                7,
+                2,
+                4,
+                5,
+                7,
+                0,
+                5,
+                7,
+                5,
+                5,
+                9,
+                5,
+                2,
+                3,
+                9,
+                7,
+                4,
+                1,
+                1,
+                1,
+                2,
+                7,
+                7,
+                7,
+                1,
+                0,
+                1,
+                4,
+                3,
+                1,
+                1,
+                3,
+                2,
+                1,
+                3,
+                5,
+                0,
+                3,
+                7,
+                7,
+                2,
+                2,
+                7,
+                7,
+                0,
+                2,
+                7,
+                1,
+                5,
+                7,
+                6,
+                5,
+                8,
+                5,
+                5,
+                1,
+                6,
+                7,
+                9,
+                6,
+            ]
+        )
+        # assert np.isclose((wawa.worker_score - np.array([1 / 2, 1, 1, 1 / 2])).sum(), 0)
+        print([e == y_ for e, y_ in zip(expected, y10)])
+        assert all([e == y_ for e, y_ in zip(expected, y10)])
 
 
 def test_twothird():
@@ -68,7 +792,7 @@ def test_twothird():
 
     sparse = [True, False]
     for sparse_ in sparse:
-        two = TwoThird(ANSWERS, n_classses=2, n_workers=4, sparse=sparse_)
+        two = TwoThird(ANSWERS, n_classes=2, n_workers=4, sparse=sparse_)
         y = two.get_answers()
         expected = np.array([1, 0, 1])
         assert all([e == y_ for e, y_ in zip(expected, y)])
